updated datatypes of input base model & added linting ,removed redundant packages

+++ 02-production-model-package/classification_model/processing/validation.py
@@ -47,7 +47,7 @@ class TitanicDisasterSurvivalDataInputSchema(BaseModel):
     Pclass: Optional[int]
     Name: Optional[str]
     Sex: Optional[str]
-    Age: Optional[int]
+    Age: Optional[float]
     SibSp: Optional[int]
     Parch: Optional[int]
     Ticket: Optional[str]
